fixed explicit data types, added workflow

+++ autograd/cpu/losses.py
@@ -70,13 +70,13 @@ class CCELoss:
         
         if not mask:
             loss             = (target * (pred.softmax_log())).sum(axis=(0,1,2))
-            loss_multiplier  = Tensor(value=np.ones_like(loss.value)*(-1/pred.shape[0]))
+            loss_multiplier  = Tensor(value=np.ones_like(loss.value)*(-1/pred.shape[0]), dtype=pred.dtype)
             loss             = loss * loss_multiplier
         else:
             mask_np          = (target[:,:,0] != 1)[...,None]
             eos_no_mask      = ((target[:,:,0] == 1).cumsum(axis=1) == 1)[...,None]
-            mask             = Tensor(mask_np | eos_no_mask, learnable=False, leaf=True)
+            mask             = Tensor(mask_np | eos_no_mask, learnable=False, leaf=True, dtype=pred.dtype)
             loss             = ((target*mask)*(pred.softmax_log())).sum(axis=(0,1,2))
-            loss_multiplier  = Tensor(value=np.ones_like(loss.value)*(-1/np.sum(mask[:])))
+            loss_multiplier  = Tensor(value=np.ones_like(loss.value)*(-1/np.sum(mask[:])), dtype=pred.dtype)
             loss             = loss * loss_multiplier
         return loss
